+++ test/foundry/unit/ClientChainGateway.t.sol
@@ -374,7 +374,7 @@ contract WithdrawalPrincipalFromExocore is SetUp {
 
         // Simulate adding VIRTUAL_STAKED_ETH_ADDRESS to whitelist via lzReceive
         bytes memory message =
-            abi.encodePacked(GatewayStorage.Action.REQUEST_ADD_WHITELIST_TOKEN, abi.encodePacked(tokens[0], uint256(0)));
+            abi.encodePacked(GatewayStorage.Action.REQUEST_ADD_WHITELIST_TOKEN, abi.encodePacked(tokens[0], uint128(0)));
         Origin memory origin = Origin({srcEid: exocoreChainId, sender: address(exocoreGateway).toBytes32(), nonce: 1});
 
         vm.prank(address(clientChainLzEndpoint));
@@ -384,7 +384,7 @@ contract WithdrawalPrincipalFromExocore is SetUp {
         origin.nonce = 2;
         message = abi.encodePacked(
             GatewayStorage.Action.REQUEST_ADD_WHITELIST_TOKEN,
-            abi.encodePacked(tokens[1], restakeToken.totalSupply() / 20)
+            abi.encodePacked(tokens[1], uint128(restakeToken.totalSupply() / 20))
         );
         vm.prank(address(clientChainLzEndpoint));
         clientGateway.lzReceive(origin, bytes32(0), message, address(0), bytes(""));
